% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGCA_call_met.R
\name{ACGCA}
\alias{ACGCA}
\title{ACGCA model function}
\usage{
ACGCA()
}
\arguments{
\item{sparms}{A vector containing the parameters for the simulationl.}

\item{gparms}{A vector containing values that control simulation behavior.}

\item{r0}{The starting radius. Defaults to 0.05.}
}
\description{
This function provides an easy interface for running the ACGCA model.
}
\keyword{IBM}
